feat(windows): Add build.bat script
This builds with the gnu toochain, relying on mingw64. And it copies
necessary .dll files.

I don't suppose there's a good way to automate the library copying? And
while gtk apparently doesn't support static linking, it would be nice if
there was some way to reduce some of the library code here...