feat(windows): Update WiX installer to handle upgrades
Running a newer .msi will uninstall the old version of the configurator
and install the new one over it.

By having WiX generate every GUID other than the UpgradeCode (which
shouldn't change), an upgrade only needs to update the Version without
messing with any of the GUIDs.

Installing the same version, for instance CI builds, is also handled.