improv: Refactor daemon code using macro to reduce redundance
Macro magic like this is potentially confusing, but arguably it's no
less confusing to have to manually add each new method in various
places. Now it only has to be added to the commands! call, and the
server.

So dark sorcery, but it should be easier to work with ultimately. (Hail
Cthulhu!)