improv: Make KeyboardColorButton a GTK widget
This makes it possible to remove a `@strong` reference, preventing a
reference cycle.

Previously there was a reference cycle because the GTK widget needed a
reference to `KeyboardColorButtonInner` for callbacks, and likewise that
struct contained a reference to `KeyboardColorButtonInner`. The
reference in the callback couldn't be weak, because in some uses it is
expected that `KeyboardColorButton` may be dropped by code using the
type, to just use the GTK widget.

This adds a bit of boilerplate, but it's probably the cleanest way to
avoid reference cycles.

The cycle here prevented the `DaemonClient`'s drop from being called.