improv: Refactor Picker to be a gtk widget, unentangled from Keyboard
By making it a Gtk widget rather than just an ordinary struct, it isn't
necessary to wrap in `Rc`, and the Rust struct will be freed when the
widget is no longer in use without causing reference cycles.

`Picker` has a strong reference to `Keyboard`, which has a weak
reference. It would be nicer if they interacted though binding glib
properties, but then it isn't type safe...

This uses `xvfb-run` to run the unit tests on CI, so `gtk::init()`
won't fail.