improv: Embed required icons in gresource
This should work on Windows, AppImage, macOS, or anything else, without
having to worry about the packaging for each individually.

Copied from Adwaita. `window-*` should be in `ui` rather than `actions`,
but because of how loading icons from resources works, that fails
because the icons are interpreted as part of HiColor, which lacks a `ui`
directory.

Yeah, Gtk... but it seems to work.