feat: Menu option to Invert F Keys
Implements https://github.com/pop-os/keyboard-configurator/issues/24.

Not sure if there's a better way to describe this, or any other way to
improve it, but it seems appropriate to provide an easy way to do this.
It is a common option provided by laptop firmware these days, and users
have complained that it's overly complicated to change each key
manually.

This has no impact on the default Launch layout since nothing is bound
on layer 2 for the F keys. Not sure if that could be confusing or could
be improved.