feat: 'Non-US \' on builtin keyboards
'Non-US #' is not listed, since the PS/2 keycode for that is the same as
'Backlash'. The UI around non-US layouts could be improved, but that's
already an issue and this at least makes it possible to set this
mapping.

This does not add the `PAUSE` and `SCROLL_LOCK` keycodes that have
been added to EC. At least with `PAUSE`, this is complicated since it
won't work on older firmware, and we should probably test for that
(https://github.com/pop-os/keyboard-configurator/issues/140).

Fixes https://github.com/pop-os/keyboard-configurator/issues/127.