feat: Extras, tap-hold, and modifier keys
Initial implementation of
https://github.com/pop-os/keyboard-configurator/issues/146.

Backend representation for mod-tap, layer-tap, and modifiers+basic
keycode, with conversion to and from text format and numeric keycode.

Add shift-select for modifiers with a non-modifier key, extras tab, and
tap-hold section.

Sensitivity needs to be set based on whether shift is held and what is
already selected. Getting this to work properly is awkward with how GTK
modifiers are handled. More things need to be added to extras, the
layout for basic keys needs to be updated for the latest design, and
various style details need to be addressed. And it needs to not show
these things for internal keycodes.