Only read keys in layout during nelson test
Other Launch keyboards used every key in the keymap, so it was
fine to read any key in the matrix. Though, on launch heavy there
are several unmapped keys, these ones obvisouly don't get
pressed by nelson, because they don't pyhsically exist.

This ignores those extra keys by only checking for pressed keys
that are in both the matrix and the layout.

squash