Pass and store `--launch-test` value to backend
The daemon thread needs to know if the keyboard configurator is in
testing mode, this will allow for modified behaviour when a
keyboard is plugged in, but before the GUI has access to the
keyboard.

The plan for this is to allow the daemon_thread to send a different
signal if the keyboard's firmware isn't completely updated.
Preventing Production from being able to test and ship keyboards
with obsolite firmware. Though this isn't added in this commit.