Use `fwupdmgr get-update --json`
Rather than trying to parse for a specific device,
`get-updates --json` will only include the name of a device
if there is a firmware update for it. If the output contains
`Configurable Keyboard` we can assume for production's sake,
that that is a `Launch * Configurable Keyboard`, and that it isn't
updated.