Require updated firmware to test (#156)
* Pass and store `--launch-test` value to backend

The daemon thread needs to know if the keyboard configurator is in
testing mode, this will allow for modified behaviour when a
keyboard is plugged in, but before the GUI has access to the
keyboard.

The plan for this is to allow the daemon_thread to send a different
signal if the keyboard's firmware isn't completely updated.
Preventing Production from being able to test and ship keyboards
with obsolite firmware. Though this isn't added in this commit.

* Don't load launch test when keyboard isn't updated

If the keyboard isn't updated do not load the configurator page.
never send the `BoardLoadingDone` signal, rather send
`BoardnotUpdated`, but only do this if the configurator is in
`--launch-test` mode.

* Run cargo fmt

* Fix cargo tests

As the tests arn't designed to test production's workflow
it is fine to pass `false` when creating a backend. This will run
the tests exactly as they used to run.

* Fix compile warnings

* Fix build for windows and mac

* Use serde "PascalCase"

* Use `fwupdmgr get-update --json`

Rather than trying to parse for a specific device,
`get-updates --json` will only include the name of a device
if there is a firmware update for it. If the output contains
`Configurable Keyboard` we can assume for production's sake,
that that is a `Launch * Configurable Keyboard`, and that it isn't
updated.

* Do not remove board if not in hashmap

Caused a crash if keyboard was removed before board was completely
initialized.