Override scancodes for boards with new firmware
Moving from version 0.7.103 to 0.19.12 of QMK changed the
scancode values. Because the keyboard configurator currently
uses scancodes to identify the keys, this creates many issues
with how each key is identified and assigned in the configurator.

This checks for the version of the keyboard and uses a different
map of scancode to scancode_names.