Update backend to use channel, and not depend on glib
The glib based version isn't easily usable in a multi-threaded context,
or outside of a glib mainloop. This works for the GTK UI, but is also
suitable for building an Iced UI on top of. It may be helpful for
implementing a command line interface too.

This changes a lot of code, but behavior should generally be unchanged.