fix(ci): Update Github Action from `ubuntu-18.04` to `ubuntu-20.04`
Apparently if you use a runner name that doesn't exist, Github Actions
doesn't tell you and just shows "Waiting for a runner to pick up this
job..." until it times out after some number of hours. Including for an
old runner that doesn't exist any more. So this should fix the CI
builds.

This may make the appimage stop working on some older distros. I doubt
many people are trying to run it on anything older than 20.04 anyway? If
this is an issue, we can do the build in a Docker container.