Use lsub to guarentee correct firmware is flashed.
Checks for USB ID of both the usb hub, and the DFU bootloader.
Using the USB IDs of the two we can identify the keybaord
plugged in, even when the keyboard doesn't say its name (which in
bootloader mode they dont).
The flashing menu is now only visible when the keyboard is in
bootloader mode as well, meaning that it is harder to accidenly
flash without knowing what one is doing, and makes the UX more
clear.