chore: Deduplicate `layouts/` and improve layout generation
Instead of having a `keymap.json` per "keyboard", and an `overrides`
directory with variants of each, there are now 3 keymap json files: `ec`,
`qmk`, and `qmk_legacy`.

This has always been a bit redundant, but `overrides` made it worse, so
this is an improvement. This moves filtering out of LED keycodes
depending on the board to `backend`, instead of `layouts.py`.

`led.json` generation still needs to be updated for the new version of
QMK.