Make `PAUSE` keycode available depending on EC build date
This should be a reliable way to test support for released firmware.
This also adjusts the default keymap to match layer 1 instead of using
the invalid key.

This should allow `PAUSE` to be handled correctly on current firmware,
without having old EC firmware allow this keycode and use it in
resetting layout, but not have it work.