windows: Fix `build.py` with latest msys2
It seems msys2 is no-longer packaging a 32-bit version of imagemagick.
(But it's only a build dependency, so that's not a problem.)

I also see it using `C:\windows\system32\convert.exe` (an unrelated
program) in some cases, even when `which` lists a different version?
Odd, unless that was just caching.

Anyway, this installs and uses the right version.